requirements.txt
telethon
python-dotenv
aiogram
